% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qswat.R
\name{check_qswat}
\alias{check_qswat}
\title{Check for QSWAT+ delineation problems and fix them by nudging outlets(s) downstream}
\usage{
check_qswat(data_dir, make_plot = TRUE, nudge_dist = NULL, check_main = FALSE)
}
\arguments{
\item{data_dir}{character path the the project directory}

\item{make_plot}{logical, whether to draw a plot showing the QSWAT+ delineation results}

\item{nudge_dist}{numeric distance (metres) to move inlets/outlets}
}
\value{
list with elements 'nudge' and 'trim'
}
\description{
This checks for inlet/outlet snapping issues in the stream network shape file of a
QSWAT+ project. If any problems are detected the function attempts to fix them by
nudging (ie re-positioning) the problematic outlet points in the downstream direction
by distance \code{nudge_dist} (m), or else suggests to increase the \code{snap_threshold} in
the \code{run_qswat} call.
}
\details{
Imprecision in elevation models and/or outlet coordinates can prevent QSWAT+
from correctly delineating your stream network. This function checks for two possible
issues and attempts to fix them automatically:
\enumerate{
\item Not all channels drain to the main outlet (ie the chain of \code{LINKNO} keys is broken somewhere)
\item Some/all of the inlet drainage remains attached to the basin (ie QSWAT+ failed to remove it)
}

Both problems can usually be resolved by re-positioning the outlet point
to fall on a different DEM cell. Users can do it by hand using the "Review Snapped"
feature in QSWAT+. This function automates the process, by nudging problematic outlets
by distance \code{nudge_dist} along a vector pointing downstream. The direction vector for
nudging connects the centroid of the outlet's (upstream) flow line with the existing
outlet point. The default distance (\code{nudge_dist=NULL}) is set to the average of the
(x and y) side lengths of the DEM grid cells.

The new positions are returned as sf data frame 'nudge' in a list along with the
affected upstream sub-basin polygons (sf data frame 'trim'). If no problems are detected,
the function returns empty data frames. If \code{check_main=TRUE} and (1) occurs, then (2) is
ignored and the function returns only the new main outlet position. If \code{check_main=FALSE}
and (1) occurs, the function ignores it and checks only (2).

When \code{make_plot=TRUE} the function draws the DEM heatmap and overlays:
\itemize{
\item the original NHDPlus model catchment boundary with inlets/outlets (black line, white circles)
\item the gage positions (if any, as grey circles)
\item the QSWAT+ sub-basins (transparent white fill and white boundaries)
\item the QSWAT+ stream reaches (blue lines)
\item problematic QSWAT+ sub-basins, if any (transparent red fill)
\item the gage positions after snapping by QSWAT+ (large black circle outlines)
\item the suggested gage positions (if any, as solid black circles with white outline)
}
}
