% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clipr}
\alias{clipr}
\title{Clip and mask a raster to a polygon}
\usage{
clipr(r, b, p = NULL)
}
\arguments{
\item{r}{SpatRaster}

\item{b}{sf polygon geometry within extent of \code{r}}

\item{p}{character path to save as GeoTIFF}
}
\value{
SpatRaster
}
\description{
This returns a copy the sub-grid of raster \code{r} covering geometry \code{b}
(for boundary), with all pixels outside of \code{b} set to NA. If \code{p} is not
\code{NULL}, the function deletes anything at that path, then writes the raster
to it with \code{terra::writeRaster} (creating directories as needed)
}
\details{
The function wraps the obvious \code{terra} functions, handling projections,
and coercing to \code{SpatVector} via \code{sp} to avoid bugs with unusual geometries.
}
