% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{comid_down}
\alias{comid_down}
\title{Return a vector of all downstream COMID values}
\usage{
comid_down(comid, edge, first_only = FALSE, from = "FROMCOMID", to = "TOCOMID")
}
\arguments{
\item{comid}{character vector, the key(s) downstream of \code{comid}}

\item{edge}{data frame, the NHD "PlusFlow" table or similar (see \code{?get_upstream})}
}
\value{
character vector, the key values encountered on the way to the outlet (in order)
}
\description{
This uses the \code{from} and \code{to} fields in the \code{edge} table (by default 'FROMCOMID'
and 'TOCOMID') to traverse the directed graph of flow lines downstream from the
input(s) \code{comid}, all the way to the eventual outlet(s). The function returns
all keys encountered along the way in a vector.
}
\details{
\code{comid} should be listed in \code{edge[['FROMCOMID']]}, otherwise the function
simply returns its first argument. \code{edge} is normally generated by \code{get_catch}
(which calls \code{get_upstream} to fetch it via \code{nhdR}) but it can be any data
frame with the same type of graph structure.
}
\examples{
# dummy table - these are not real COMIDs!
edge_eg = data.frame(
 FROMCOMID=c('00', '00', '10', '11', '20'),
 TOCOMID=c('10', '11', '20', '20', '30'))

# follow from upstream of fork
comid_down('00', edge_eg)

# follow from one side of fork then the other
comid_down('10', edge_eg)
comid_down('11', edge_eg)

}
