% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{comid_up}
\alias{comid_up}
\title{Return a vector of all upstream COMID values}
\usage{
comid_up(comid, edge, first_only = FALSE)
}
\arguments{
\item{comid}{character vector, the COMID values to start from}

\item{edge}{data frame, the NHD "PlusFlow" table (see \code{?get_upstream})}
}
\value{
character vector, the COMID values for polygons draining into the outlet
}
\description{
This uses the 'FROMCOMID' and 'TOCOMID' fields in the \code{edge} table to follow
the directed graph of flow lines upstream from the polygon(s) represented by
the input \code{comid}. The function exhaustively checks all relevant branches in
a loop, returning all upstream COMID values in a vector
}
\details{
All entries of \code{comid} should be listed in \code{edge[['FROMCOMID']]}. Normal usage
is to pass a single value to \code{comid} identifying your outlet of interest. The
function uses vectorized calls internally.

\code{edge} is normally generated by \code{get_catch} (which calls \code{get_upstream} to fetch
it via \code{nhdR}) but it can be any data frame with the same type of graph structure
in the fields 'FROMCOMID' and 'TOCOMID' (see examples).
}
\examples{
# dummy table - these are not real COMIDs!
edge_eg = data.frame(
 FROMCOMID=c('00', '00', '10', '11', '20'),
 TOCOMID=c('10', '11', '20', '20', '30'))

# follow from downstream of fork to get all nodes
comid_up('30', edge_eg)

# follow from one side of fork then the other
comid_up('10', edge_eg)
comid_up('11', edge_eg)
}
