% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwis.R
\name{data_nwis}
\alias{data_nwis}
\title{Download daily observation data for a particular NWIS site and variable}
\usage{
data_nwis(
  nwis,
  site,
  from = NULL,
  param_code = "00060",
  stat_code = "00003",
  n_sec = 0.5
)
}
\arguments{
\item{nwis}{data frame returned by \code{dataRetrieval::whatNWISdata}}

\item{site}{character, the site number (must appear in \code{site_no} column of \code{nwis})}

\item{from}{Date, the earliest date to fetch}

\item{param_code}{character, the 5-digit NWIS parameter code (see \code{?usgs_lookup})}

\item{stat_code}{character the 5-digit statistic code for the variable (default is mean)}

\item{n_sec}{numeric >= 0, the number of seconds to wait between requests}
}
\value{
data frame with 'date' column and the requested values
}
\description{
This uses the result of \code{dataRetrieval::whatNWISdata} to construct API calls
for waterservices.usgs.gov/rest to download daily records.
}
\details{
This processes one variable at a time. Specify the variable code in \code{param_code},
along with the site code \code{site}, and the starting date \code{from}. If \code{from} is not
supplied, the function downloads all available dates.

It is important to respect rate limits when making requests to this API.
The default \code{n_sec} sets a half-second pause in between subsequent requests,
when the site has multiple distinct records. You should program a similar delay
to happen in between calls to \code{data_nwis} (for example using \code{Sys.sleep}).
}
