% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{datum_to_epsg}
\alias{datum_to_epsg}
\title{EPSG codes for some common geographical coordinate systems}
\usage{
datum_to_epsg(code)
}
\arguments{
\item{code}{character, the datum name}
}
\value{
integer, the EPSG code
}
\description{
These are some of the datum names I have encountered in NWIS. If \code{code}
is not 'NAD27' or 'NAD83', the function assumes the WGS84 datum.
}
\examples{
datum_to_epsg('WGS84')
datum_to_epsg('NAD27')
datum_to_epsg('some_unknown_datum')
}
