% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{draw_outlet}
\alias{draw_outlet}
\title{Draw one or more points with filled circles}
\usage{
draw_outlet(p, cex = 1, col_in = "white", col_out = "grey20")
}
\arguments{
\item{p}{any POINT object understood by \code{sf::st_geometry()}}

\item{cex}{passed to \code{plot}}

\item{col_in}{character color of the border}

\item{col_out}{character fill color}
}
\value{
nothing but draws on existing plot
}
\description{
Wrapper for \code{sf::plot.sf(p)}. This adds to an existing plot in the same
coordinate system as \code{p}.
}
