% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{draw_scale}
\alias{draw_scale}
\title{Title something}
\usage{
draw_scale(
  obj,
  bottom = TRUE,
  left = TRUE,
  size = 0.5,
  draw = TRUE,
  y_adj = 0,
  outer_adj = 0.05
)
}
\arguments{
\item{obj}{the sf object that was used to create the plot}

\item{bottom}{logical whether to position the scale bar at top or bottom}

\item{left}{logical logical whether to position the scale bar at left or right}

\item{size}{numeric > 0 controls the width of the scale bar (see details)}

\item{draw}{logical}

\item{y_adj}{numeric > 1}
}
\value{
something
}
\description{
Title something
}
