% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_all.R
\name{fetch_all}
\alias{fetch_all}
\title{Fetches all required data for a QSWAT+ model build}
\usage{
fetch_all(
  outlet,
  data_dir,
  overwrite = FALSE,
  force_overwrite = FALSE,
  nwis_nm = "flow_ft",
  nwis_from = as.Date("2005-01-01"),
  no_download = FALSE
)
}
\arguments{
\item{outlet}{geo-referenced point object passed to \code{sf::st_geometry}, the main outlet}

\item{data_dir}{character path to the directory to use for output files}

\item{overwrite}{logical, should the function create missing files in \code{data_dir}?}

\item{force_overwrite}{logical, should the function overwrite existing files in \code{data_dir}?}

\item{no_download}{logical, if TRUE \code{nhdR::nhd_plus_get} calls is skipped (for debugging)}
}
\value{
a list of file paths to write
}
\description{
By default (\code{overwrite=FALSE}) this returns a list of the file paths written by the
function, in the order they are processed. With \code{overwrite=TRUE} the function generates
each group of files in sequence by downloading data (from USGS and USDA) on various
upstream catchment features.
}
\details{
The function produces a set of files in \code{data_dir} sufficient for building a SWAT+ model
for the basin draining to \code{outlet}. Output files are ready to use with QSWAT+.

Steps can be run individually, and if any fails you should be able to try again by
deleting to sub-directory for that step.
}
