% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catch.R
\name{get_catch}
\alias{get_catch}
\title{Return a list of NHD geometry objects modelling the catchment for an outlet}
\usage{
get_catch(outlet, crs_out = 4326, fast = FALSE, no_download = FALSE)
}
\arguments{
\item{outlet}{character to query OSM for location or else a point accepted by \code{sf::st_geometry}}

\item{crs_out}{output CRS, anything acceptable to \code{sf::st_crs} or NULL to use UTM zone of outlet}

\item{fast}{logical, if TRUE some detailed objects are omitted from results, speeding things somewhat}

\item{no_download}{logical, if TRUE the \code{nhdR::nhd_plus_get} call is skipped (for debugging)}
}
\value{
list
}
\description{
Download and extract geometries and other data related the NHDPlus model for the
catchment of \code{outlet}, using \code{nhdR}. If \code{outlet} is a character string, it
is first passed to \code{nominatim_point} to get coordinates from OpenStreetMap.
}
\details{
The function uses \code{nhdR::nhd_plus_get} to fetch the relevant data packages, setting
\code{temporary=FALSE} so that persistent storage is used to cache files instead of
downloading them repeatedly between R sessions. In subsequent calls to \code{get_catch},
you can set \code{no_download=TRUE} to avoid \code{nhdR::nhd_plus_get} altogether (this can
be helpful if you the data already cached locally but are getting 404 errors).

The function returns a list with the following
\itemize{
\item \code{comid} integer key value associated with the outlet
\item \code{outlet} sf points data frame, the outlet point (snapped to NHDPlus model) and some metadata
\item \code{boundary} sfc_POLYGON, the catchment boundary inscribing all areas draining to \code{outlet}
}

If \code{fast=FALSE}, the returned list also includes
\itemize{
\item \code{edge} data frame of edges in the directed graph representing stream flow in the NHD model
\item \code{catchment} sf data frame with polygon geometries, the upstream (sub)catchments
\item \code{flow} sf data frame with line geometries, the upstream stream reaches (possibly artificial)
\item \code{lake} sf data frame with polygon geometries, the upstream water bodies
}

All geo-referenced outputs are converted to the common projection \code{crs_out}. If this
is \code{NULL} (the default) the function uses the UTM zone overlying \code{outlet}.

See also \code{?get_upstream} which handles computations for this function.

Note that in the NHDPlus model the landscape is partitioned into small non-overlapping
patches of land, each draining to a unique outlet point. The function finds the polygon
on which the user-supplied \code{outlet} lies, then snaps the outlet to polygon's drainage point -
ie it returns the outlet associated with the COMID, which may be slightly downstream of the
input coordinates. See https://cran.r-project.org/package=nhdplusTools and the references
pages linked there for more on the NHDPlus data model, and how the COMID ties things together.
}
