% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catchment.R
\name{get_catchment}
\alias{get_catchment}
\title{Return a list of geometry objects describing watershed upstream}
\usage{
get_catchment(outlet, crs_out = NULL)
}
\arguments{
\item{outlet}{character to query OSM for location or else a point accepted by \code{sf::st_geometry}}

\item{crs_out}{output CRS, anything acceptable to \code{sf::st_crs} or NULL to use UTM zone of outlet}
}
\value{
list
}
\description{
This uses the \code{nhdR} package to manage downloads of NHD data collections
related to the watershed upstream of the point \code{outlet}.  If \code{outlet} is
a character string instead, it is passed to \code{nominatim_point} to get the point
from OSM.
}
\details{
The function downloads three collections (see \code{?get_upstream}) from NHD
containing highly detailed information on sub-catchments and stream reaches,
which is opened and processed to determine the catchment for \code{outlet}.

Note that \code{nhdR} uses \code{rappdirs} to manage persistent storage. This will set
up a default data storage folder on your system and use cached files instead
of downloading repeatedly whenever possible.

The function returns a list with the following
}
