% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_land.R
\name{get_land}
\alias{get_land}
\title{Wrapper for \code{FedData::get_nlcd}}
\usage{
get_land(data_dir)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}
}
\value{
SpatRaster, land use data for the catchment area (direct from \code{FedData::get_nlcd})
}
\description{
This loads the bounding box for the DEM created by \code{get_dem} and passes it
to \code{FedData::get_nlcd}, which handles downloading and ETL. The result is returned
as a SpatRaster rather than a RasterLayer.
}
