% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil.R
\name{get_soil}
\alias{get_soil}
\title{Soil getter}
\usage{
get_soil(data_dir)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}
}
\value{
something
}
\description{
Soil getter
}
