% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soils.R
\name{get_statsgo}
\alias{get_statsgo}
\title{Fetch STATSGO2 or SSURGO data}
\usage{
get_statsgo(data_dir, model = "statsgo")
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{model}{character either 'statsgo' or 'ssurgo'}
}
\value{
sf dataframe object containing a "MUKEY" field and polygon geometries
}
\description{
Downloads and opens STATSGO2 or SSURGO data covering the bounding box of the DEM
created by \code{get_dem}. For STATSGO2 we use HTTPS to download from the FPAC Box cloud
system at the USDA (see ?statsgo_url). For SSURGO we use \code{FedData::get_ssurgo} to
download SSA tile archives and handle ETL with \code{sf} and \code{terra}.
}
\details{
The function saves its source files to the 'raw' sub-directory of the 'statsgo' and
'ssurgo' sub-directories of \code{data_dir}. Get the path to this storage directory for the
SSURGO model by calling the following:

\code{save_statsgo(data_dir, model='ssurgo', overwrite=FALSE)[['raw_dir']]}

SSURGO is very detailed, so when using this model the memory usage may be high,
and the intermediate .gpkg file written to the "raw" sub-directory may be large.
For example with the Carter's Bridge example (Upper Yellowstone) the memory usage
exceeds 10 GB and about 2 GB are written to disk.
}
