% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_utm}
\alias{get_utm}
\title{Find the appropriate UTM zone for a geometry}
\usage{
get_utm(x, quiet = FALSE)
}
\arguments{
\item{x}{any object with CRS data understood by \code{sf::st_geometry}}
}
\value{
integer EPSG code
}
\description{
This prints the UTM zone that covers the supplied point
and returns the corresponding EPSG code as integer
}
\examples{
c(-110, 45) |> sf::st_point() |> get_utm()
}
