% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{land_use_lookup}
\alias{land_use_lookup}
\title{A lookup table matching NLCD and SWAT+ land use codes}
\format{
\subsection{\code{land_use_lookup}}{

A data frame with 272 rows and 3 columns:
\describe{
\item{id}{integer NLCD ID}
\item{name}{SWAT+ land use code }
\item{description}{character string with both matched descriptions (NLCD on the right)}
}
}
}
\source{
\url{https://doi.org/10.5194/hess-22-2527-2018}
}
\usage{
land_use_lookup
}
\description{
This data set matches the National Land Cover Database (NLCD) product
from the Multi-Resolution Land Characteristics (MRLC) consortium
to SWAT+ land use codes. Matches are found using Table A.3 from
the Hydrol. Earth Syst. Sci paper by Havel, Tasdighi, Arabi (2018).
}
\details{
SWAT+ land use codes were extracted from the SQLite
database that is included in the default SWAT+ installer
("swatplus_datasets.sqlite"). Since the SWAT+ team updates these
codes from time to time, this dataset will likely need updating
someday. It was last generated in June 2023.

If a SWAT+ code cannot be matched using the table, it is assigned
the name \code{NA}. Note that the paper by Havel, Tasdighi and Arabi
focused on a mountainous basin similar (and close) to the UYR
watershed. Other regions of interest will likely need a different
set of mappings reflecting the local vegetation.

Learn more about SWAT+ land use codes in the SWAT+ documentation PDF
at https://swat.tamu.edu/media/69419/Appendix-A.pdf
}
\keyword{datasets}
