% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwis.R
\name{list_nwis}
\alias{list_nwis}
\title{Fetch records of point data from stations at the National Water Information System (NWIS)}
\usage{
list_nwis(
  data_dir,
  nwis_nm = "flow_ft",
  param_code = "00060",
  stat_code = "00003",
  n_min = 10,
  buff_m = 100
)
}
\arguments{
\item{data_dir}{character path to the output files directory}

\item{nwis_nm}{character name of the output data sub-directory in "nwis"}

\item{param_code}{character, the 5-digit NWIS parameter code (see \code{?usgs_lookup})}

\item{stat_code}{character the statistic code for the variable (default is mean)}

\item{n_min}{integer > 0, the minimum number of records to include a station}

\item{buff_m}{numeric > 0, the length in metres to buffer the catchment boundary (see details)}
}
\value{
list with data frames 'catch' and 'all'
}
\description{
This calls \code{dataRetrieval::whatNWISdata} to get a list of available service records
for the area of interest. This is defined by the catchment boundary file at
\code{save_catch(data_dir)['boundary']}, extended outwards by \code{buff_m} metres.
}
\details{
Currently this only checks for daily values (having \code{data_type_cd='dv'}). Records with
fewer than \code{n_min} observations are skipped.

The function retuns a list with two data frames: 'all' is the output from
\code{dataRetrieval::whatNWISdata} before filtering for observation count and variable name;
and 'catch' is an sf points data frame describing the station sites of interest.

Specify the statistic of interest by its stat code \code{stat_code} (see \code{?points_nwis}).
Specify the variables of interest by supplying \code{param_code}. A lookup table for these
codes is lazy-loaded by this package in object \code{usgs_lookup} - check the definition
of your query by calling:

\code{usgs_lookup |> dplyr::filter(parm_cd == param_code)}

For more information on accessing NWIS from R see the help pages and vignettes at
https://cran.r-project.org/package=dataRetrieval
}
