% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wx.R
\name{load_wx}
\alias{load_wx}
\title{Get weather data for a SWAT+ model from \code{wxArchive}}
\usage{
load_wx(
  wx_dir,
  data_dir,
  sub_dir = NULL,
  var_nm = "pcp_mean",
  subbasin_id = NULL,
  from = NULL,
  quiet = FALSE
)
}
\arguments{
\item{wx_dir}{character path to the data directory for \code{wxArchive}}

\item{data_dir}{character path to the \code{rqswat} data directory for the (sub)catchment}

\item{var_nm}{character, one of "tmp_max", "tmp_min", "hum_mean", "pcp_mean", "wnd_mean"}

\item{subbasin_id}{integer key from QSWAT+ identifying the sub-basin}

\item{from}{integer or \code{NULL}, the year to start from (this and all later years copied)}
}
\value{
list with 'values' (the data), 'date' (mapping to rows), 'poly' (mapping to columns)
}
\description{
This is a helper function for \code{wxArchive} users. It returns a data frame of
daily weather data for sub-basin(s) in the QSWAT+ model in \code{data_dir}. Note
that this requires having an up-to-date \code{wxArchive} project whose extent
covers the requested sub-basin(s).
}
\details{
Before calling this function, users must first save the output of
\code{load_qswat(data_dir, 'sub', sub=TRUE)} to "aoi_export.geojson" in \code{wx_dir},
then run the "downscale" and "extract" operations in \code{wxArchive} to compute
the time series. See \code{?wxArchive::workflow_downscale} and
\code{?wxArchive::workflow_extract} for details.

Optional argument \code{subbasin_id} specifies the QSWAT+ "Subbasin" code of a
particular set of sub-basins. Leave this at its default \code{NULL} to return data
for all sub-basins.

The five options for \code{var_nm} are designed to serve as drop-in replacements for
the standard SWAT+ weather station data files - eg values from \code{var_nm='pcp_mean'}
can be copied to ".pcp" files using \code{rswat_open} and \code{rswat_write}.
}
