% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qswat.R
\name{make_weather}
\alias{make_weather}
\title{Create a set of dummy weather files positioned at sub-basin centroids}
\usage{
make_weather(
  data_dir,
  overwrite = FALSE,
  var_nm = NULL,
  to = NULL,
  from = NULL,
  pts = NULL
)
}
\arguments{
\item{data_dir}{character, path to the directory for input/output files}

\item{overwrite}{logical, whether to write the output or just return the file paths}

\item{var_nm}{character vector, variable names to write}

\item{to}{Date, final date in the daily time series}

\item{from}{Date, initial date in the daily time series}

\item{pts}{sf data frame, with points geometry and 'Subbasin' key (integer column)}
}
\value{
vector of file paths
}
\description{
This writes empty weather files for the specified date range - ie all values are
NA, which triggers the SWAT+ weather generator.
}
\details{
A weather station is created at the centroid of each sub-basin from the QSWAT+ project
in \code{data_dir} (or at \code{pts}, if supplied), and the associated files are written to
the "TxtInOut" directory for the project. NAs are assigned for all data values (to be
replaced later by the user).

\code{pts} is used internally to avoid opening the sub-basin shape file repeatedly.
It normally is used to pass the sub-basin centroids, but any set of points can be
passed as long as they have a coherent set of 'Subbasin' keys (ie one for each of
the project's sub-basins keys), and 'Elevation' values (meters)
}
