% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{nominatim_point}
\alias{nominatim_point}
\title{Return a point given a place-name query using Nominatim API for OpenStreetMaps (OSM)}
\usage{
nominatim_point(query)
}
\arguments{
\item{query}{character, query to submit to Nominatim API}
}
\value{
sfc_POINT object or NULL for no results
}
\description{
This submits a single query to OSM and returns either a single result, the
WGS84 coordinates for the point that (most closely) matches the query, or else
\code{NULL} to indicate no results.
}
\details{
The function also prints the "display_name" field returned by OSM

Loosely based on code by D. Kisler at https://tinyurl.com/2xxcthp5
See also the Nominatim API documentation at
https://nominatim.org/release-docs/latest/api
}
\examples{
nominatim_point('nonexistent placename')
nominatim_point('Old Faithful')
}
