% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catch.R
\name{open_catch}
\alias{open_catch}
\title{Open catchment model geometries saved by \code{save_catch}}
\usage{
open_catch(data_dir, extra = NULL)
}
\arguments{
\item{data_dir}{character path to the output files directory}

\item{extra}{logical, if \code{TRUE} the function includes additional files used with `get_split}
}
\value{
a list with elements 'edge', 'outlet', 'catchment', 'flow', 'lake', 'boundary'
}
\description{
With default \code{extra=NULL} the function loads the extra files only if they all exist.
The function will otherwise proceed with some/all files missing and simply omit the
missing object (with a warning) or, in the case of \code{edge}, returning a \code{NULL} in its place
}
\details{
See \code{get_catch} for more detail on the output elements
}
