% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_catch}
\alias{plot_catch}
\title{Plot the catchment for an outlet}
\usage{
plot_catch(catch_list, main = NULL, crs_out = NULL)
}
\arguments{
\item{catch_list}{list returned by \code{get_catch}}

\item{main}{character, a title for the plot}
}
\value{
invisibly returns the scale bar line geometry from \code{draw_scale}
}
\description{
This plots the catchment area, flow lines, and lakes/ponds upstream of
the \code{outlet} passed to \code{get_catch}, along with the outlet itself (circled).
}
\details{
The function adds a scale bar and sets the title automatically (as needed)
by copying the most frequently used name in the 'GNIS_NAME' field of the flow
lines dataset and appending the COMID for the catchment polygon overlying the
outlet.

All input geometries are transformed to coordinates in \code{crs_in} for plotting.
By default this is set to UTM zone of the main outlet.
}
