% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwis.R
\name{points_nwis}
\alias{points_nwis}
\title{Summarize output from \code{dataRetrieval::whatNWISdata} as a sf points data frame}
\usage{
points_nwis(nwis_info, param_code = NULL, stat_code = "00003", na_rm = TRUE)
}
\arguments{
\item{nwis_info}{subset of data frame returned by \code{dataRetrieval::whatNWISdata}}

\item{param_code}{character vector, parameter code(s) (see \code{?usgs_lookup})}

\item{stat_code}{character the 5-digit statistic code for the variable (default is mean)}

\item{na_rm}{logical if \code{TRUE}, sites with no applicable records are omitted}
}
\value{
data frame with a row for each distinct station site and three columns
}
\description{
The function returns an sf data frame summarizing the NWIS metadata returned by
\code{dataRetrieval::whatNWISdata} on hydrological point data. Points are in WGS84
coordinates and fields include site metadata and data availability.
}
\details{
This reshapes the data frame to have one row per site. Since a site may
produce multiple records, the function adds a field (column) for each of the
requested variables (in \code{param_code}) indicating the latest date for which a
record is available at the site (or \code{NA} if no records are available).

\code{param_code} can be \code{NULL} to return info on all available variables. It it is
specified and its entries are named, these names are re-used in the output.
For example if \code{param_code=c(flow_m='30208')} then the output data frame will
include a column named \code{flow_m} with the latest available date (or \code{NA}).
Otherwise the \code{param_code} itself is the column name.

The \code{count} field in the output is the sum of the counts for all of the requested
variables - ie if you request one variable, then \code{count} is the number of days of
records, but if you request multiple variables \code{count} is the sum of these numbers.

The 5-digit \code{stat_code} by default fetches mean values ('00003'). This makes
sense for our use-case of daily stream flow data, but not necessarily for other
applications. Other options include '00011' (instantaneous), '00001' (max),
'00002' (min).

Sites with none of the requested variables are omitted from the results with
the default \code{na_rm=TRUE}. Set \code{na_rm=FALSE} to keep them.

See also also the help pages at https://help.waterdata.usgs.gov/
}
