% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qswat.R
\name{qswat_edge}
\alias{qswat_edge}
\title{Make a data frame of channel network links from QSWAT+ channels shape-file}
\usage{
qswat_edge(swat_channels)
}
\arguments{
\item{swat_channels}{sf data frame with LINESTRING geometries, created by QSWAT+}
}
\value{
data frame with integer columns 'FROMCOMID', 'TOCOMID'
}
\description{
This creates a data frame of channel linkages based on the three link number fields
in the channels shape file returned by QSWAT+ (and/or TauDEM). The output should be
compatible with functions from this package having an \code{edge} argument, like \code{comid_up}
\code{comid_down} and \code{find_outlet}.
}
\details{
This helper function allows us to re-use functions designed for NHD data. It constructs
the fields 'FROMCOMID' and 'TOCOMID' based on the integer 'LINKNO', 'USLINKNO1', 'USLINKNO2'
fields in \code{swat_channels}.

Note that 'FROMCOMID' and 'TOCOMID' are so named for compatibility reasons and have
nothing to do with the NHD COMID system. They simply link together the rows of \code{edge}
as a directed graph. Relate this back to your QSWAT+ project via the 'LINKNO' key.

Get \code{swat_channels} by first calling \code{run_qswat} to run QSWAT+, then find the shape file
path in the JSON \code{run_qswat(data_dir)[['output']]} and pass it to \code{sf::st_read}.
}
