% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qswat.R
\name{report_moved}
\alias{report_moved}
\title{Check if any inlet/outlet points were moved this package and/or QSWAT+ and report the distances}
\usage{
report_moved(
  data_dir,
  draw = FALSE,
  line_col = "black",
  snapped = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{data_dir}{character path the the project directory}
}
\value{
the latest outlet points data frame with new field 'distance' (m)
}
\description{
This reports the total distances between different versions of outlet points
after re-positioning by \code{check_qswat} and saving the results with a \code{run_qswat} call.
}
\details{
With default arguments \code{run_qswat} will run \code{check_qswat} and then re-run QSWAT+ setup
until the project passes the checks, each time changing the path of the outlets file
to a new modified version. This function checks if this has happened, and if so it
reports the distances
}
