% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wx.R
\name{save_aoi}
\alias{save_aoi}
\title{Save sub-basins file TODO}
\usage{
save_aoi(data_dir, name = "aoi_export", overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{data_dir}{}

\item{name}{}

\item{overwwrite}{}
}
\description{
Save sub-basins file TODO
}
