% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catchment.R
\name{save_catchment}
\alias{save_catchment}
\title{Save the output of \code{get_catchment} to disk}
\usage{
save_catchment(data_dir, catch_list = NULL, overwrite = FALSE)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{catch_list}{list returned from \code{get_catchment(..., fast=FALSE)}}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}
}
\value{
the file names to write
}
\description{
When \code{overwrite=TRUE} the function writes 'outlet.geojson', 'catchment.geojson',
'flow.geojson', 'lake.geojson', and 'boundary.geojson' (by passing the like-named
objects to \code{sf::st_write}), and when \code{overwrite=FALSE} the function writes nothing
but returns the file paths that would be written.
}
\details{
The outlet file contains the COMID as a field. All outputs are in WGS84 coordinates.
See \code{get_catchment} and \code{get_upstream} for details on input datasets
}
\examples{
save_catchment('/example')
}
