% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_land.R
\name{save_land}
\alias{save_land}
\title{Save the output of \code{get_land} to disk}
\usage{
save_land(data_dir, land = NULL, overwrite = FALSE)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{land}{list returned from \code{get_catch(..., fast=FALSE)}}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}
}
\value{
the file names to write
}
\description{
Save the output of \code{get_land} to disk
}
\examples{
save_land('/example')
}
