% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwis.R
\name{save_nwis}
\alias{save_nwis}
\title{Save the output of \code{get_nwis} to disk}
\usage{
save_nwis(data_dir, nwis = NULL, overwrite = FALSE)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{nwis}{list returned from \code{get_nwis}}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}
}
\value{
the file names to write
}
\description{
The function writes a copy of \code{nwis} to geoJSON. The contents of the "raw"
sub-directory are written by \code{get_nwis}.
}
\details{
When \code{overwrite=TRUE} the function writes ..., and when \code{overwrite=FALSE} the function
writes nothing but returns the file paths that would be written.
}
\examples{
save_nwis('/example')
}
