% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qswat.R
\name{save_qswat}
\alias{save_qswat}
\title{Make a QSWAT+-friendly copy of the output of \code{get_catch} or \code{get_split}}
\usage{
save_qswat(
  data_dir,
  sub = FALSE,
  overwrite = FALSE,
  lake_area = 0.5,
  quiet = FALSE
)
}
\arguments{
\item{data_dir}{character path to the directory for input/output files}

\item{sub}{logical, if \code{TRUE} the function processes sub-catchments in "split/"}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}

\item{lake_area}{numeric (in km2), lakes with area lower than \code{lake_area} are omitted}
}
\value{
the file names to write
}
\description{
This function creates a set of files ready to load in QSWAT+ and sufficient to
define and construct a basic SWAT+ model. This includes: the DEM, land use, and
soil MUKEY rasters (as GeoTIFF); a lookup table for land use; an inlets/outlets
shape file; and a stream network shape file (to burn in the DEM).
}
\details{
When \code{overwrite=TRUE} the function writes these files to the "qswat" sub-directory
of the catchment specified by \code{data_dir}. If \code{sub=TRUE} the function does this in a
loop over the sub-catchment found in the 'split' sub-directory of \code{data_dir}. When
\code{overwrite=FALSE} the function writes nothing but returns the file paths that would
be written.

All geo-referenced outputs are written in UTM coordinates, where the zone is
determined by the main outlet point location (see \code{?to_utm}). Lakes are burned into
the land use raster (code 'watr') and all raster outputs are cropped and masked to
the catchment boundary.

Note that weather files are the responsibility of the user, and should be added after
calling this function (and possibly after running QSWAT+ to find HRU locations).
}
\examples{
save_qswat('/example')
}
