% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil.R
\name{save_soil}
\alias{save_soil}
\title{Save the output of \code{get_soil} to disk}
\usage{
save_soil(data_dir, statsgo = NULL, ssurgo = NULL, overwrite = FALSE)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{overwrite}{logical}
}
\value{
the file names to write
}
\description{
Save the output of \code{get_soil} to disk
}
\examples{
save_soil('/example')
}
