% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_split.R
\name{save_split}
\alias{save_split}
\title{Save results of \code{get_split}}
\usage{
save_split(
  data_dir,
  sub_list = NULL,
  overwrite = FALSE,
  nwis_nm = "flow_ft",
  param_code = "00060",
  stat_code = "00003"
)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{sub_list}{list returned from \code{get_split}}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}

\item{nwis_nm}{character, passed to \code{nwis_split}}

\item{param_code}{character, passed to \code{nwis_split}}

\item{stat_code}{character, passed to \code{nwis_split}}

\item{buffer}{}
}
\value{
list of two: the gage file path and a vector of directory paths
}
\description{
This saves the sub-catchments returned by \code{get_split} to sub-directories of
\code{file.path(data_dir, 'split')} with directory names copied from \code{names(sub_list)}.
The final outputs from \code{get_dem}, \code{get_land}, \code{get_soil}, and \code{get_nwis} are then
filtered to each sub-catchment and written to disk in a loop. A single points data
frame is also written to 'split/gage.geojson', mapping gages to sub-directories.
}
\details{
All raster files are cropped and masked to the union of the sub-catchment boundary
and the inlet NHD polygon(s) (if any. See \code{?get_split}). Note that \code{gage}

The function returns a list with: \code{gage}, the path to the mapping file; and \code{sub},
a vector of paths to the sub-catchment directories. Pipe any of these directories to
\code{save_catch}, \code{save_dem}, \code{save_land}, \code{save_soil}, or \code{save_nwis} to get a list of the
files saved in each category.

Once 'split/gage.geojson' has been written by this function, subsequent calls to
\code{update_nwis(data_dir)} will distribute copies of relevant NWIS files to the
sub-catchment directories of "split" so that each one is individually also up to date.

\code{param_code} and \code{stat_code} select the stream gage variable to copy from the NWIS
dataset for the catchment, and \code{nwis_nm} is a name for it. These arguments have no effect
when listing the (parent) directory paths with \code{overwrite=FALSE}.
}
\examples{
save_split('/example')
save_split('/example', extra=TRUE)
}
