% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wx.R
\name{save_wx}
\alias{save_wx}
\title{Write weather data from \code{load_wx} to SWAT+ weather station data files}
\usage{
save_wx(sub_dir, wx_list = NULL, overwrite = FALSE, quiet = FALSE)
}
\arguments{
\item{sub_dir}{character path to the (sub)catchment directory}

\item{wx_list}{list returned from \code{load_wx} for \code{sub_dir}}

\item{overwrite}{logical, if FALSE the function returns the file paths but writes nothing}

\item{quiet}{logical, suppresses console messages}
}
\value{
vector of file paths modified
}
\description{
Write weather data from \code{load_wx} to SWAT+ weather station data files
}
