% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qswat.R
\name{split_catch}
\alias{split_catch}
\title{Split a catchment into sub-catchments with outlets on the points in \code{gage}}
\usage{
split_catch(gage, edge, catchment)
}
\arguments{
\item{gage}{sf points data frame, point of interest at which to split the catchment}

\item{edge}{data frame of flow line directed edges ("PlusFlow" from "NHDPlusAttributes")}

\item{catchment}{sf data frame of sub-catchment polygons ("Catchment" from "NHDPlusCatchment")}
}
\value{
list with elements 'boundary', 'gage', 'gage_lookup'
}
\description{
This finds a partition of the union of the polygons in \code{catchment} based on the
locations of points of interest in \code{gage}, and the directed edges representing flow
lines in \code{edge}. The output has one sub-catchment for each point in \code{gage}, with
\code{gage} positioned on the outlet polygon.
}
\details{
Arguments \code{edge} and \code{catchment} can be generated by passing a main outlet point to
\code{get_upstream} (which uses \code{nhdR::nhd_plus_load} to fetch and load the data). The
function expects \code{gage} to include both this main outlet and at least one other point.

For each point in \code{gage}, the function finds the overlying polygon in \code{catchment},
then merges it with all upstream polygons to form a larger polygon representing the
drainage of the point. If a different \code{gage} point lies upstream, then its drainage
polygon is differenced from the result to form the sub-catchment polygon.

This is done recursively starting from the main outlet, resulting in a set of disjoint
polygons ("sub-catchments") covering the entire catchment. Each has a unique outlet
in \code{gage}, and possibly one or more inlets (the upstream elements of \code{gage}).

The output list has two elements
\itemize{
\item \code{boundary} is an sf data frame of sub-catchment polygons and flow directions
\item \code{gage_lookup} is a data frame mapping points from \code{gage} to polygons in \code{boundary}
}

Sub-catchments in \code{boundary} are assigned identifiers from their associated \code{gage}
('station_nm' and 'comid'). If multiple \code{gage} points lie on the same \code{catchment}
polygon, then they will share a common sub-catchment polygon in \code{boundary}. If this
happens, the function copies the name and COMID from the station with the highest
'count' field (ie the most records).
}
