% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwis.R
\name{split_nwis}
\alias{split_nwis}
\title{Distribute copies of NWIS data files to the relevant sub-catchments}
\usage{
split_nwis(data_dir, nwis_nm, param_code = "00060", stat_code = "00003")
}
\arguments{
\item{data_dir}{character path to the output files directory}

\item{nwis_nm}{character name of the output data sub-directory in "nwis"}

\item{param_code}{character, the parameter code to update}

\item{stat_code}{character, the statistic code to update}
}
\value{
character vector, paths to the directories modified
}
\description{
This is a helper for the \code{get_split} workflow, where sub-catchment data is saved
to sub-directories of \code{file.path(data_dir, 'split')}. This function copies data from
gages located at/near the outlet of each catchment, as well as any inlets, to the 'nwis'
sub-directory for that catchment.
}
\details{
The function returns a vector of paths to the directories modified. Once the
function has been called, users can pass any of these sub-directories to
\code{update_nwis} to update only the gages associated with that catchment, or
pass \code{data_dir} to update them all.
}
