% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statsgo_url}
\alias{statsgo_url}
\title{A table of HTTP links to download STATSGO2 data}
\format{
\subsection{\code{statsgo_url}}{

A data frame with 20 rows and 5 columns:
\describe{
\item{id}{FPAC Box ID}
\item{file}{the zip archive file name}
\item{abb}{state/territory abbreviation code}
\item{state}{state or NA for territories}
\item{url}{URL to download the file with HTTPS}
}
}
}
\source{
\url{https://nrcs.app.box.com/v/soils/folder/18247487156}
}
\usage{
statsgo_url
}
\description{
The State Soil Geographic (STATSGO2) dataset is an older source of soils data
with a different coverage than the newer (and more detailed) SSURGO product
from the USDA. This table provides links to download zipped STATSGO2 soil maps
for all 51 US states and three territories.
}
\details{
These links are to a publicly accessible FPAC Box that I found by Google search.
The filenames suggest they were last modified in 2016 and uploaded in 2017. This
suits our purposes, but users may want to check if newer versions are available.

Learn more about SSURGO/STATSGO2 on the NRCS metadata links page at
https://www.nrcs.usda.gov/resources/data-and-reports/ssurgo/stats2go-metadata
}
\keyword{datasets}
