% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swatplus_soils}
\alias{swatplus_soils}
\title{A vector of STATSGO and SSURGO soil MUKEYs available in SWAT+}
\format{
\subsection{\code{swatplus_soils}}{

A vector with:
\describe{
\item{abbr}{abbrevation}
\item{full}{state name}
\item{geometry}{POLYGON with WGS84 coordinates}
}
}
}
\source{
\url{https://swat.tamu.edu/media/69419/Appendix-A.pdf}
}
\usage{
swatplus_soils
}
\description{
SWAT+ MUKEYs were extracted from the SQLite database that is included in the default
SWAT+ installer ("swatplus_datasets.sqlite"). Since the SWAT+ team updates these
codes from time to time, this dataset will likely need updating someday. It was
last generated in June 2023.
}
\keyword{datasets}
