% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to_utm}
\alias{to_utm}
\title{Find the appropriate UTM zone for a geometry}
\usage{
to_utm(x)
}
\arguments{
\item{x}{any object understood by \code{sf::st_geometry}}
}
\value{
integer EPSG code
}
\description{
This prints the UTM zone that covers the centroid of the supplied geometry
and returns the corresponding EPSG code
}
\examples{
c(-110, 45) |> sf::st_point() |> to_utm()
}
