% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwis.R
\name{update_nwis}
\alias{update_nwis}
\title{Update}
\usage{
update_nwis(data_dir, from = NULL, param_code = "00060", stat_code = "00003")
}
\arguments{
\item{data_dir}{}

\item{from}{}

\item{param_code}{}

\item{stat_code}{}
}
\description{
Update
}
