% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_states}
\alias{us_states}
\title{An sf object containing USA state polygons and their names}
\format{
\subsection{\code{us_states}}{

An sf data frame with 51 rows and 3 columns:
\describe{
\item{abbr}{abbrevation}
\item{full}{state name}
\item{geometry}{POLYGON with WGS84 coordinates}
}
}
}
\source{
\url{https://cran.r-project.org/package=us_map}
}
\usage{
us_states
}
\description{
Data come from the \code{usmaps::us_map('state', ...)} function. This dataset provides
a version as an sf data frame object with MULTIPOLYGON geometry and WGS84 coordinates.
}
\keyword{datasets}
