% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usgs_lookup}
\alias{usgs_lookup}
\title{A table of USGS parameter code definitions}
\format{
\subsection{\code{usgs_lookup}}{

A data frame with 24953 rows and 13 columns:
\describe{
\item{parm_cd}{5-digit parameter code as character}
\item{group}{category of variable}
\item{parm_nm}{definition}
\item{parm_unit}{unit}
...
}
}
}
\source{
\url{https://help.waterdata.usgs.gov/}
}
\usage{
usgs_lookup
}
\description{
Parameter codes are used by the US Geological Survey (USGS) National Water
Information System (NWIS) to unambiguously identify variables in published
datasets. This table provides name/definition pairs for these codes.
}
\details{
This data set comprises the name/definition pairs (along with some other fields)
found at the following URL:

'https://help.waterdata.usgs.gov/code/parameter_cd_query?fmt=rdb&group_cd=\%'

This was last retrieved in June 2023.

More information about the parameter codes can be found at the USGS NWIS Help site:
https://help.waterdata.usgs.gov/codes-and-parameters/parameters
}
\keyword{datasets}
