% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wx.R
\name{write_wx}
\alias{write_wx}
\title{Write weather data from \code{load_wx} to SWAT+ weather station data files}
\usage{
write_wx(
  sub_dir,
  wx_list,
  overwrite = FALSE,
  bounds = NULL,
  add = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{sub_dir}{character path to the (sub)catchment directory}

\item{wx_list}{list returned from \code{load_wx} for \code{sub_dir}}

\item{overwrite}{logical, if FALSE the function returns the file paths but writes nothing}

\item{bounds}{named list of numeric (length-2) vectors to replace default bounds}

\item{add}{logical, if \code{TRUE} the function keeps existing rows for dates not found in \code{wx_list}}

\item{quiet}{logical, suppresses console messages}
}
\value{
vector of file paths modified
}
\description{
Unfinished
}
