% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{biggest_poly}
\alias{biggest_poly}
\title{Omit all but the largest in a set of polygons}
\usage{
biggest_poly(poly_list)
}
\arguments{
\item{poly_list}{any object for which \code{sf::st_geometry() |> sf::st_area()} is valid}
}
\value{
an sfc_polygon object
}
\description{
This selects the largest polygon from a set, transforms to WGS84 coordinates,
runs \code{sf::st_make_valid()}, then removes holes from the result.
}
\details{
This is useful for tidying up the result of set operations (like differencing)
when the expected result is a single connected polygon.
}
