% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_all.R, R/get_all.R
\name{fetch_all}
\alias{fetch_all}
\title{Fetches all required data for a QSWAT+ model build}
\usage{
fetch_all(data_dir, outlet = NULL, overwrite = FALSE, force_overwrite = FALSE)

fetch_all(data_dir, outlet = NULL, overwrite = FALSE, force_overwrite = FALSE)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{outlet}{point object accepted by \code{sf::st_geometry}}

\item{overwrite}{logical, should the function create missing files in \code{data_dir}?}

\item{force_overwrite}{logical, should the function overwrite existing files in \code{data_dir}?}
}
\value{
a list of file paths to write

a list of file paths to write
}
\description{
Fetches all required data for a QSWAT+ model build

Fetches all required data for a QSWAT+ model build
}
