% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{fetch_all}
\alias{fetch_all}
\title{Download input data and run QSWAT+ for the basin defined by \code{outlet}}
\usage{
fetch_all(
  outlet,
  data_dir,
  overwrite = FALSE,
  force_overwrite = FALSE,
  nwis_nm = "flow_ft",
  nwis_from = as.Date("2005-01-01"),
  no_download = FALSE
)
}
\arguments{
\item{outlet}{geo-referenced point object passed to \code{sf::st_geometry}, the main outlet}

\item{data_dir}{character path to the directory to use for output files}

\item{overwrite}{logical, should the function create missing files in \code{data_dir}?}

\item{force_overwrite}{logical, should the function overwrite existing files in \code{data_dir}?}

\item{no_download}{logical, if TRUE \code{nhdR::nhd_plus_get} calls is skipped (for debugging)}
}
\value{
a list of file paths to write
}
\description{
By default (\code{overwrite=FALSE}) this returns a list of the file paths written by the
function, in the order they are processed. When \code{overwrite=TRUE} the function runs
the model setup workflow on the catchment upstream of \code{oulet}, by downloading and
processing watershed data and passing it to QSWAT+.
}
\details{
With default settings and \code{overwrite=TRUE}, the function does the following:
\enumerate{
\item downloads public data on the landscape and hydrology
\item partitions the data into sub-catchments that can be modelled individually
\item runs automated QSWAT+ setup on each of the sub-catchments
\item runs SWAT+ Editor to build "TxtInOut" from the QSWAT+ project database
}

For details on (1) see \code{?get_catch}, \code{?get_nwis}, \code{?get_dem}, \code{?get_land}, \code{?get_soil},
\code{?get_split} (which are called in sequence); For details on partitioning, see \code{?get_split};
And for details on SWAT+ setup (3-4), see \code{?save_qswat} and \code{?run_qswat}.

"TxtInOut" is the directory name for a SWAT+ model, ie for the set of plaintext
configuration files that parametrize the simulator. The path to this directory can
be passed to \code{rswat} to manage model fitting and execution.

Each step in this workflow produces a new sub-directory in \code{data_dir}, and for every
\verb{get_*} function there is a \verb{save_*} function that returns the file paths written within:
eg for \code{get_dem}, call \code{save_dem(data_dir)} to get the paths. Users can repeat a step
manually for debugging by deleting the sub-directory for the step and running the
\verb{get_*(...) |> save_*(...,overwrite=TRUE)} pair again.
}
