% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{get_dem}
\alias{get_dem}
\title{Wrapper for \code{FedData::get_ned}}
\usage{
get_dem(data_dir, pad_size = NULL)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{pad_size}{units object, the width of padding (or NULL to set default)}
}
\value{
SpatRaster of elevation data for the catchment area
}
\description{
This opens the boundary polygon created by \code{get_catchment} and passes (a
buffered version of) it to \code{FedData::get_ned}.
}
\details{
The buffer size default is 10\% of the square root of the polygon area. For
a square domain this is equal to 1/10 of the side length.

Note that \code{FedData} downloads tiles to a temporary folder so I don't think
we can cache them easily for later use. Save a copy with \code{save_dem}
}
