% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwis.R
\name{get_nwis}
\alias{get_nwis}
\title{Fetch records of point data from stations at the National Water Information System (NWIS)}
\usage{
get_nwis(
  data_dir,
  param_code = c(flow_m = "30208", flow_ft = "00060"),
  n_min = 10
)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{param_code}{character, the 5-digit NWIS parameter code (see \code{?usgs_lookup})}

\item{n_min}{integer > 0, the minimum number of records to include a station}

\item{force_overwrite}{logical, causes the function to download a fresh table of service records}
}
\value{
list of data frame
}
\description{
This calls \code{dataRetrieval::whatNWISdata} to get a list of available service records,
before calling \code{nwis_data} in a loop to download all records relevant to the query.
}
\details{
Currently this only retrieves daily values (having \code{data_type_cd='dv'}) but
other types of records may be listed in nwis_info_all. Records with fewer than n_min
observations are skipped.s

Specify your variables of interest by supplying a vector of parameter codes in
\code{param_code}. A lookup table for these codes is lazy-loaded by this package in object
\code{usgs_lookup} - get a table of definitions for your query by calling:

\code{usgs_lookup |> dplyr::filter(parm_cd \%in\% param_code)}

For more information on accessing NWIS from R see the help pages and vignettes at
https://cran.r-project.org/package=dataRetrieval
}
