% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwis.R
\name{get_nwis}
\alias{get_nwis}
\title{Create/update a database of NWIS daily mean values}
\usage{
get_nwis(
  data_dir,
  nwis_nm = "flow_ft",
  param_code = "00060",
  stat_code = "00003",
  n_min = 10,
  from_initial = NULL,
  refresh = FALSE
)
}
\arguments{
\item{data_dir}{character path to the output files directory}

\item{nwis_nm}{character name of the output data sub-directory in "nwis"}

\item{param_code}{character, the parameter code to update}

\item{stat_code}{character, the statistic code to update}

\item{n_min}{integer > 0, the minimum number of observations to include a record}

\item{from_initial}{Date, results are filtered to records occurring on or after this date}

\item{refresh}{logical if TRUE a fresh list of station metadata is requested from NWIS}
}
\value{
character vector, the file paths of the output
}
\description{
This uses the \code{dataRetrieval} package to manage downloads of data from the
National Water Information Service (NWIS). By default this constructs and/or
updates a database of mean daily stream flow in ft3.
}
\details{
Station sites of interest are selected by filtering results to points within
the catchment boundary polygon created by \code{get_catch}. A GeoJSON file locating
these points is written to a subdirectory of "nwis/". To get the full path do
\code{save_nwis(data_dir, nwis_nm)}. This file is written once only when \code{refresh=FALSE}.
Set \code{refresh=TRUE} to create it again - note that this updates all fields and checks
for new station locations relevant to the whole query.

By default records from all time periods are returned. If \code{from} is set, the
function filters results to stations having a record on/after \code{from}.

Outputs from individual stations are cached in the "raw" sub-directory of the
\code{nwis_nm} directory in \code{data_dir}, where they are updated/appended as needed.
After initializing, subsequent calls will update starting from the last recorded
date in the "raw" output files. To start from scratch, delete these files.

\code{from_initial} and \code{n_min} apply only to selecting relevant stations, so it
affects the output when initializing a new database, or when updating station
info with \code{refresh=TRUE}, but not when running an ordinary daily values update
with \code{refresh=FALSE}.

Specify a different variable of interest by changing the parameter and
statistic codes \code{param_code} and  \code{stat_code} (see also \code{?load_nwis}, and
\code{?dataRetrieval::whatNWISdata}). The default settings select mean daily stream
flow in ft3/s.

When changing \code{from_initial}, \code{param_code}, or \code{stat_code}, you should either
change/delete the output directory (\code{nwis_nm}), or else set \code{refresh=TRUE} to
rebuild the stations file - otherwise the function will only be aware of
previously seen stations. Note that the default \code{nwis_nm} only makes sense for
the default \code{param_code}. The function will not warn about nonsensical name choices!
}
