% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_split.R
\name{get_split}
\alias{get_split}
\title{Split a catchment into sub-catchments with an NWIS gage at each outlet}
\usage{
get_split(data_dir, gage = NULL, snap_main = units::set_units(100, m))
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{gage}{sf points data frame, point of interest at which to split the catchment}

\item{snap_main}{numeric with units, snapping distance to set "main" outlet}
}
\value{
a list with elements 'gage' and 'sub'
}
\description{
This partitions the catchment in \code{data_dir} into a set of sub-catchments
matching the layout of upstream NWIS gages, returning metadata about the
gage points in \code{gage} and a nested list of geometries in \code{sub}, one for
each sub-catchment.
}
\details{
The inlets and outlets linking sub-catchments are placed on the same NHD
catchment polygons as the \code{gage} points. This ensures that each gage is
associated with the outlet for a single sub-catchment, and that the inlets
for this sub-catchment (if any) are themselves situated at \code{gage} points.

Sub-catchments inherit name and ID fields from row of \code{gage} mapping to its
outlet. If multiple elements in \code{gage} map to the same sub-catchment, the one
with the highest 'count' field (most records) is used, and the others discarded
when building outlet objects.

If the main outlet of the catchment is not found within snapping distance
\code{snap_main} of a point in \code{gage}, the function appends the outlet point
automatically. This ensures the set of output sub-catchment polygons always
forms a partition of the whole catchment.

The "true" outlet point for a sub-catchment (under the NHD model) lies at the
intersection of the boundary of the NHD polygon for the outlet COMID, and its
flow line. These points are calculated and returned in \code{outlet}. Note that \code{gage}
points from NWIS will often be located slightly upstream of the true outlet for
their NHD polygon. Thus two sets of boundary polygons are returned: \code{boundary}
is the partition; and \code{boundary_outer} is a copy where the polygons for any
inlets have been joined to the downstream sub-catchment boundaries.

See also \code{split_catch}, which does most of the work of following flow-lines and
building boundaries with set operations.
}
