% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil.R
\name{get_statsgo}
\alias{get_statsgo}
\title{Fetch STATSGO2 data}
\usage{
get_statsgo(data_dir, s2 = FALSE)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{s2}{logical, controls whether to use spherical approximation}
}
\value{
sf dataframe object containing a "MUKEY" field and polygon geometries
}
\description{
Downloads and opens STATSGO2 data covering the bounding box of the DEM
created by \code{get_dem}. This uses HTTPS to download from the FPAC Box cloud
system at the USDA (see ?statsgo_url).
}
\details{
The function saves its source files to the 'raw' sub-directory. Get the
path to this storage directory with \code{save_statsgo(data_dir, overwrite=FALSE)[['raw_dir']]}
}
