% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwis.R
\name{load_nwis}
\alias{load_nwis}
\title{Load NWIS data from a file on disk, or return the date range within, or the file path}
\usage{
load_nwis(
  site,
  data_dir,
  nwis_nm = "flow_ft",
  output = "path",
  param_code = "00060",
  stat_code = "00003"
)
}
\arguments{
\item{site}{character, the site code to fetch}

\item{data_dir}{character path to the output files directory}

\item{nwis_nm}{character name of the output data sub-directory in "nwis"}

\item{output}{character, either 'path', 'date' , or 'values'}

\item{param_code}{character, the parameter code to update}

\item{stat_code}{character, the statistic code to update}
}
\value{
either the file path (character), date range (length-2 POSIXct vector), or data frame
}
\description{
This function organizes the storage of bulk data downloads from NWIS. The data are
stored in CSV files in the "raw" sub-directory of \code{nwis_nm}. File names are constructed
by joining the site, parameter and statistic codes (in that order) separated by an
underscore.
}
\details{
With the default \code{output} set to 'path', the function returns the expected file path for
the CSV. With 'date', the function attempts to open the file and returns its date range
or NULL. With 'data' the function returns the whole contents of the file as a data frame.
}
