% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_qswat.R
\name{load_qswat}
\alias{load_qswat}
\title{Load shape files from a QSWAT+ project}
\usage{
load_qswat(data_dir, quiet = FALSE)
}
\arguments{
\item{data_dir}{character path to "qswat" subdirectory}
}
\value{
list with geometries 'sub', 'channel', 'outlet' and parameter 'nudge_default'
}
\description{
This loads the channels, sub-basins, and outlets shape-files created
during QSWAT+ setup and returns them as sf geometry data frames, in a list
along with a suggested nudge distance for moving outlets (see \code{?check_qswat}).
}
\details{
This is meant for projects created using \code{run_qswat}; The function expects an
input JSON file in the "qswat" directory (parent of the QGIS project directory)
which it uses to find the correct input DEM path. The output JSON in this directory
(if found) is used to locate the output  QSWAT+ "Shapes" directory. If QSWAT+
completes delineation but halts afterwards (eg due to failed checks or database
issues), the function  will attempt load the shape files anyway by guessing their
path in the \code{data_dir} directory tree.

The function returns only the subset of sub-basin polygons with positive 'Subbasin'
keys, and the channels mapping to them. The returned outlets have coordinates snapped
to channels already (by QSWAT+).
}
