% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wx.R
\name{load_wx}
\alias{load_wx}
\title{Get weather data for a SWAT+ model from \code{wxArchive}}
\usage{
load_wx(
  wx_dir,
  data_dir,
  var_nm = NULL,
  sub_dir = NULL,
  subbasin_id = NULL,
  from = NULL,
  check_fields = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{wx_dir}{character path to the data directory for \code{wxArchive}}

\item{data_dir}{character path to the \code{run_maker} data directory for the (sub)catchment}

\item{var_nm}{character, see choices by calling with \code{var_nm=NULL}}

\item{sub_dir}{character, the sub-catchment directory name}

\item{subbasin_id}{integer key from QSWAT+ identifying the sub-basin}

\item{from}{integer or \code{NULL}, the year to start from (this and all later years copied)}

\item{check_fields}{logical, for internal use}

\item{quiet}{logical, suppresses console messages}
}
\value{
list with 'values' (the data), 'date' (mapping to rows), 'poly' (mapping to columns)
}
\description{
This is a helper function for \code{wxArchive} users. It returns a data frame of
daily weather data for sub-basin(s) in the QSWAT+ model in \code{data_dir}. Note
that this requires having an up-to-date \code{wxArchive} project whose extent
covers the requested sub-basin(s).
}
\details{
This function simply loads existing output from \code{wxArchive} in a convenient
format. Before calling this function, users must call \code{save_aoi} to write
"aoi_export.geojson" in \code{wx_dir}, then run the "downscale" and "extract"
operations in \code{wxArchive} to compute the output time series. See
\code{?wxArchive::workflow_downscale} and \code{?wxArchive::workflow_extract} for details.

Optional argument \code{subbasin_id} specifies the QSWAT+ "Subbasin" code of a
particular set of sub-basins. Leave this at its default \code{NULL} to return data
for all sub-basins.

The options for \code{var_nm} are designed to serve as drop-in replacements for
the standard SWAT+ weather station data files - eg values from \code{var_nm='pcp_mean'}
can be copied to ".pcp" files using \code{rswat_open} and \code{rswat_write}.
}
