% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qswat.R
\name{make_weather}
\alias{make_weather}
\title{Create a set of dummy weather files positioned at sub-basin centroids}
\usage{
make_weather(
  data_dir,
  overwrite = FALSE,
  var_nm = NULL,
  to = NULL,
  from = NULL,
  pts = NULL
)
}
\arguments{
\item{data_dir}{character, path to the directory for input/output files}

\item{overwrite}{logical, whether to write the output or just return the file paths}

\item{var_nm}{character vector, variable names to write}

\item{to}{Date, final date in the daily time series}

\item{from}{Date, initial date in the daily time series}

\item{pts}{sf data frame, with points geometry and 'Subbasin' key (integer column)}
}
\value{
vector of file paths
}
\description{
This writes empty weather files for the specified date range. All data values are
set to -99. If you have weather generators set up for the project, this will cause
SWAT+ to generate (ie randomly draw) weather for the period \code{from} - \code{to} during
simulations.
}
\details{
A weather station is created at the centroid of each sub-basin from the QSWAT+ project
in \code{data_dir} (or at \code{pts}, if supplied), and a corresponding data file containing the
date range \code{from} - \code{to} in daily steps is written to  the "weather" sub-directory of
the project root.

If either \code{from} or \code{to} is missing, the function sets a 7 day period to start/end at
the supplied date. If both are missing, \code{to} is assigned \code{Sys.Date()}.

\code{pts} is used internally to avoid opening the sub-basin shape file many times in a loop.
We use it to pass the sub-basin centroids, but any set of points can be supplied
(overriding what's in \code{data_dir}) as long as they have a coherent 'Elevation' (m),
and a (1-indexed) 'Subbasin' field mapping to the project's sub-basins keys.
}
