% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{most_frequent}
\alias{most_frequent}
\title{Return the most frequent string a column of a data frame}
\usage{
most_frequent(df_nm, nm = 1, alt = "unnamed")
}
\arguments{
\item{nm}{character or integer identifying the column to check}

\item{alt}{character, returned when in place of \code{NA}}

\item{x}{data frame with character column at \code{nm}}
}
\value{
character
}
\description{
Counts appearances of unique strings in \code{x[[nm]]} and returns the most frequent
one. If there is a tie, the function returns the one appearing first in the data
frame. If there are no non-empty strings, the function returns \code{alt}
}
