% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_qswat}
\alias{plot_qswat}
\title{Plot QSWAT+ geometries}
\usage{
plot_qswat(data_dir, check_result = NULL, what = "dem", quiet = FALSE)
}
\arguments{
\item{data_dir}{character path to the data directory}

\item{check_result}{list, the result of \code{check_qswat} (to overlay)}

\item{what}{character, either 'dem', 'land', or, 'soil'}
}
\value{
nothing but draws a plot
}
\description{
Plot QSWAT+ geometries
}
