% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_plot.R
\name{plot_qswat}
\alias{plot_qswat}
\title{Plot QSWAT+ geometries}
\usage{
plot_qswat(
  data_dir,
  check_result = NULL,
  what = NULL,
  quiet = FALSE,
  add = FALSE
)
}
\arguments{
\item{data_dir}{character path to the data directory}

\item{check_result}{list, the result of \code{check_qswat} (to overlay)}

\item{what}{character, either a color name (like "black"), or one of 'dem', 'land', or, 'soil'}
}
\value{
nothing but draws a plot
}
\description{
Plot QSWAT+ geometries
}
