% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qswat.R
\name{run_editor}
\alias{run_editor}
\title{Run SWAT+ Editor to create SWAT+ simulator config files in "TxtInOut"}
\usage{
run_editor(data_dir, overwrite = FALSE)
}
\value{
vector of file paths
}
\description{
Run SWAT+ Editor to create SWAT+ simulator config files in "TxtInOut"
}
