% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_qswat.R
\name{run_qswat}
\alias{run_qswat}
\title{Run QSWAT+ with default settings to create a SWAT+ project for the catchment}
\usage{
run_qswat(
  data_dir,
  overwrite = FALSE,
  osgeo_dir = NULL,
  lake_threshold = 50L,
  channel_threshold = 0.001,
  stream_threshold = 0.01
)
}
\arguments{
\item{data_dir}{character, path to the directory for input/output files}

\item{overwrite}{logical, whether to write the output or just return the file paths}

\item{osgeo_dir}{character, path to the QGIS 3 installation directory (AKA "OSGEO4W_ROOT")}

\item{lake_threshold}{integer > 0, the percent overlap for a cell to become (part of) a lake}

\item{channel_threshold}{numeric > 0, channel creation threshold as fraction of basin area}

\item{stream_threshold}{numeric > 0, stream creation threshold as fraction of basin area}
}
\value{
list of paths related to the created QSWAT+ project
}
\description{
This executes a minimal QSWAT+ workflow (delineation, HRUs, etc) using the input files
created by \code{save_qswat}. This is Windows-only and has some external dependencies
\itemize{
\item see below. Note that when \code{overwrite=TRUE},  any existing files in the 'qswat/project/'
sub-directory of \code{data_dir} will be deleted.
}
}
\details{
All output is written to the sub-directory "qswat": QSWAT+/SWAT+ files go in the
"project" directory tree, and a set of essential input and output paths for the project
are written to "qswatplus_input.json" and "qswatplus_output.json". The function returns
the contents of "qswatplus_output.json" as an R list:
\itemize{
\item exe: path to the SWAT+ Editor executable (needed for creating the files in "TxtInOut")
\item sql: path to the project SQLite database file
\item sub: path to the project sub-basins shape file (relevant features have \code{Subbasin != 0})
\item txt: path to the project "TxtInOut" directory (configuration files for SWAT+ simulator)
}

Before calling this function you will need to install QGIS 3.32.0 and the latest SWAT+
bundle, including QSWAT+ and SWAT+ Editor. You will also need to set \code{osgeo_dir} to point
to your QGIS installation directory. By default this is set to 'C:/Program Files/QGIS 3.32.0',
which should be the correct path for many (but not all) Windows users. Unfortunately I don't
know an easy way to discover this path automatically from within R.

The function calls a Windows batch file that launches Python3 for QGIS3 and runs the
script 'run_qswatplus.py'. This script uses the QSWAT+ plugin to complete the essential
SWAT+ model creation steps, including: delineation with TauDEM, loading soils and land use
to define HRUs using "Dominant HRU" method, and creating a database to link everything.
A second function call to SWAT+ Editor produces plaintext SWAT+ config files (in "TxtInOut")
}
