% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{save_dem}
\alias{save_dem}
\title{Save the output of \code{get_dem} to disk}
\usage{
save_dem(data_dir, dem = NULL, overwrite = FALSE)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{overwrite}{logical}

\item{catch_list}{list returned from \code{get_catchment(..., fast=FALSE)}}
}
\value{
the file names to write
}
\description{
When \code{overwrite=TRUE} the function writes 'outlet.geojson', 'catchment.geojson',
'flow.geojson', 'lake.geojson', and 'boundary.geojson' (by passing the like-named
objects to \code{sf::st_write}), and when \code{overwrite=FALSE} the function writes nothing
but returns the file paths that would be written.
}
\details{
The outlet file contains the COMID as a field. All outputs are in WGS84 coordinates.
See \code{get_catchment} and \code{get_upstream} for details on input datasets
}
\examples{
save_dem('/example')
}
