% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{save_dem}
\alias{save_dem}
\title{Save the output of \code{get_dem} to disk}
\usage{
save_dem(
  data_dir,
  dem = NULL,
  overwrite = FALSE,
  res = 90,
  buffer = res,
  burn = 50,
  threads = TRUE
)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}

\item{res}{integer > 0, the resolution in metres (point spacing in both directions)}

\item{buffer}{numeric > 0, padding distance in metres (for masking to boundary)}

\item{burn}{numeric >= 0, burn-in depth for flow lines in meters}

\item{threads}{logical, passed to \code{terra::project}}

\item{catch_list}{list returned from \code{get_catch(..., fast=FALSE)}}
}
\value{
the file names to write
}
\description{
When \code{overwrite=TRUE} the function writes a bounding box polygon, 'bbox.geojson', and three
copies of the DEM: 'ned_dem.tif', the unchanged source DEM; 'dem.tif', a version warped
to the UTM zone of the main outlet (by bilinear averaging) with resolution \code{res} X \code{res};
and 'dem_burn.tif', a copy the UTM raster with stream reaches burned to depth \code{burn} (meters).
}
\details{
Burn-in refers to artificially reducing the elevation in the DEM under known stream reaches.
This is to assist the TauDEM algorithm (used in QSWAT+) in finding the correct routing
network based on the DEM alone. The function creates stream reaches by expanding flow lines
to form channel polygons of width \code{buffer} (the pixel width by default). It then uses
\code{terra::rasterizes} to reduce the value of any DEM pixel that overlaps with a channel by
the fixed value \code{burn}.

The function selects an output UTM zone for the two (non-source) rasters based on the location
of the main outlet point. The bounding box is always in WGS84 coordinates.

Note that \code{terra::project} can be slow with large rasters. The default \code{threads=TRUE} is
the faster option. Higher \code{res} will also speed computation, but results in less detail in
(SWAT+) HRUs and stream networks, and less precision in land use and soil type assignment.
}
\examples{
save_dem('/example')
}
