% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{save_dem}
\alias{save_dem}
\title{Save the output of \code{get_dem} to disk}
\usage{
save_dem(
  data_dir,
  dem = NULL,
  overwrite = FALSE,
  res = 90,
  buffer = 1000,
  threads = TRUE
)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}

\item{res}{integer > 0, the resolution in metres (point spacing in both directions)}

\item{buffer}{numeric > 0, padding distance in metres (for masking to boundary)}

\item{threads}{logical, passed to \code{terra::project}}

\item{catch_list}{list returned from \code{get_catch(..., fast=FALSE)}}
}
\value{
the file names to write
}
\description{
When \code{overwrite=TRUE} the function writes a bounding box polygon, 'bbox.geojson', and two
copies of the DEM raster.
}
\details{
The function writes the following rasters
\itemize{
\item 'ned_dem.tif', the unchanged source DEM
\item 'dem.tif', a version that has been masked/cropped to the catchment and warped to UTM
}

The function selects the UTM zone based on the location of the main outlet point. The output
grid has resolution \code{res} X `res, and warping is by nearest neighbour - this is to better
preserve extreme values (vs bilinear averaging).

The bounding box is always saved in WGS84 coordinates (the expected system for geoJSON).

Masking is controlled by \code{buffer}. Increase it to add more padding around the catchment
boundary, or set it to \code{Inf} to disable masking altogether. Note that the QSWAT+ manual
recommends against masking unless you are sure that your catchment boundary is accurate
at the scale of the DEM. In that case, masking should improve the speed and precision of
delineation with TauDEM; but otherwise it can lead to inaccurate channel networks and errors.

\code{terra::project} (which uses GDAL) can be slow with large rasters. The default \code{threads=TRUE}
is the faster option. Higher \code{res} will also speed computation, but results in less detail in
(SWAT+) HRUs and stream networks, less precision in land use and soil type assignment
down the line, and a higher risk of failure in the QSWAT+ channel delineation.
}
\examples{
save_dem('/example')
}
