% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dem.R
\name{save_dem}
\alias{save_dem}
\title{Save the output of \code{get_dem} to disk}
\usage{
save_dem(
  data_dir,
  dem = NULL,
  overwrite = FALSE,
  res = 90,
  buffer = Inf,
  threads = TRUE
)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}

\item{res}{integer > 0, the resolution in metres (point spacing in both directions)}

\item{buffer}{numeric > 0, padding distance in metres, for masking}

\item{threads}{logical, passed to \code{terra::project}}

\item{catch_list}{list returned from \code{get_catch(..., fast=FALSE)}}
}
\value{
the file names to write
}
\description{
When \code{overwrite=TRUE} the function writes a bounding box polygon and two
copies of the DEM raster. When \code{overwrite=FALSE} the function writes nothing
but returns the file paths that would be written.
}
\details{
The function writes its output to the 'ned' sub-directory of \code{data_dir}:
\itemize{
\item 'bbox.geojson', bounding box is always saved in WGS84 coordinates
\item 'ned_dem.tif', the unchanged source DEM
\item 'dem.tif', a version warped to UTM (and optionally masked/cropped to the catchment)
}

The function selects the UTM zone based on the location of the main outlet point. The UTM
grid has resolution \code{res} X \code{res}, and warping is by nearest neighbour to better preserve
extreme values (vs bilinear averaging).

Masking is controlled by \code{buffer}. Set it to zero to mask and crop to the catchment
boundary. Increase it to extend this boundary outwards, or set it to \code{Inf} not mask
or crop (the default).

See the QSWAT+ manual for a discussion of masking pros/cons.

\code{terra::project}, which uses GDAL, can be slow with large rasters. The default \code{threads=TRUE}
is the faster option. Higher \code{res} will also speed computation, but results in less detail in
(SWAT+) HRUs and stream networks, less precision in land use and soil type assignment
down the line, and a higher risk of failure in the QSWAT+ channel delineation.
}
\examples{
save_dem('/example')
}
