% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwis.R
\name{save_nwis}
\alias{save_nwis}
\title{Save NWIS files to disk}
\usage{
save_nwis(nwis_dir, nwis_list = NULL, station_data = NULL, overwrite = FALSE)
}
\arguments{
\item{nwis_dir}{character path to the directory to use for output files}

\item{nwis_list}{list returned from \code{list_nwis}}

\item{station_data}{data frame of station data (eg the result of \code{data_nwis})}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}
}
\value{
the file names to write
}
\description{
This writes the output of \code{list_nwis} (to geoJSON and CSV) and \code{data_nwis} (to CSV),
or reports the files that would be written.
}
\details{
When \code{overwrite=TRUE} the function writes non-NULL objects passed to \code{nwis_list} and
\code{station_data} to disk. When \code{overwrite=FALSE} (the default) the function writes nothing
but returns the file paths that would be written.
}
\examples{
save_nwis('/example')
}
