% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwis.R
\name{save_nwis}
\alias{save_nwis}
\title{Save NWIS files to disk}
\usage{
save_nwis(data_dir, nwis_nm = "flow_ft", nwis_list = NULL, overwrite = FALSE)
}
\arguments{
\item{data_dir}{character path to the output files directory}

\item{nwis_nm}{character name of the output data sub-directory in "nwis"}

\item{nwis_list}{list returned from \code{list_nwis}}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}
}
\value{
the file names to write
}
\description{
This writes the output of \code{list_nwis} (to geoJSON) or reports the files that would be written.
}
\details{
When \code{overwrite=TRUE} the function writes non-NULL objects in \code{nwis_list} to disk. When
\code{overwrite=FALSE} (the default) the function writes nothing but returns the file paths
that would be written.

All files written by this function go in the directory \code{data_dir |> file.path('nwis', nwis_nm)}.
Note that these are point location files, whereas the data values themselves are stored in the
"raw" subdirectory (and they are written by a different function, \code{update_nwis})
}
\examples{
save_nwis('/example')
}
