% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qswat.R
\name{save_qswat}
\alias{save_qswat}
\title{Make a QSWAT+-friendly copy of your (sub-)catchment data files}
\usage{
save_qswat(
  data_dir,
  sub = FALSE,
  overwrite = FALSE,
  quiet = FALSE,
  lake_area = 0.5,
  burn = 50,
  snap_tol = 300
)
}
\arguments{
\item{data_dir}{character path to the directory for input/output files}

\item{sub}{logical, if \code{TRUE} the function processes sub-catchments in "split/"}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}

\item{lake_area}{numeric (in km2), lakes with area lower than \code{lake_area} are omitted}

\item{burn}{numeric >= 0, stream reach burn-in depth and width, in meters}

\item{snap_tol}{numeric >= 0, snapping tolerance distance, in meters}
}
\value{
the file names to write
}
\description{
This function creates a set of files ready to load in QSWAT+ and sufficient to
define and construct a basic SWAT+ model. When \code{overwrite=FALSE} the function
writes nothing but returns the file paths that would be written.
}
\details{
When \code{overwrite=TRUE} and \code{sub=FALSE} the following are written to the "qswat"
sub-directory of your \code{data_dir}:
\itemize{
\item two copies of the DEM, with and without burn-in (as GeoTIFF)
\item a land use raster, and a STATSGO2/SSURGO MUKEY raster (as GeoTIFF)
\item a lookup table for land use (as CSV)
\item an inlets/outlets shape file
\item a stream network shape file
}

'dem_burn.tif', a copy of the UTM raster with stream reaches burned to depth \code{burn} (in meters)
Burn-in refers to artificially reducing the elevation in the DEM under known stream reaches.
This is to assist the TauDEM algorithm (used in QSWAT+) in finding the correct routing
network based on the DEM alone. The function creates stream reaches by expanding flow lines
to form channel polygons of width \code{burn}. It then uses \code{terra::rasterizes} to reduce the value
of any DEM pixel that overlaps with a channel by the fixed value \code{burn}.

Outlets and inlets are automatically snapped to the nearest flow-line whenever they lie
within \code{snap_tol} meters. If no such flow line exists, the function throws an error.

If \code{sub=TRUE} the function writes its output to the sub-catchments directories
created by \code{get_split}, in a loop. This produces a complete set of QSWAT+ files
for each sub-catchment in "split".

All geo-referenced outputs are written in UTM coordinates, where the zone is
determined by the main outlet point location (see \code{?to_utm}). Lakes are burned into
the land use raster (code 'watr') and all raster outputs are cropped and masked to
the catchment boundary.

Note that weather files are the responsibility of the user, and should be added after
calling this function (and possibly after running QSWAT+ to find HRU locations).
}
\examples{
save_qswat('/example')
}
