% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soil.R
\name{save_soil}
\alias{save_soil}
\title{Save the output of \code{get_soil} to disk}
\usage{
save_soil(
  data_dir,
  soil = NULL,
  overwrite = FALSE,
  buffer = Inf,
  threads = TRUE
)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{overwrite}{logical if \code{FALSE} the function just returns the file that would be written}

\item{buffer}{numeric > 0, padding distance in metres, for masking}

\item{threads}{logical, passed to \code{terra::project}}

\item{soils}{SpatRaster of soil MUKey values}
}
\value{
the file names to write
}
\description{
When \code{overwrite=TRUE} the function writes two copies of the soil MUKEY rasters.
When \code{overwrite=FALSE} the function writes nothing but returns a list containing
the file paths that would be written (in element 'soil') along with file paths
writting by \code{get_soil} (in elents 'ssurgo' and 'statsgo').
}
\details{
This writes output to the 'soils' sub-directory of \code{data_dir}:
\itemize{
\item 'soil_src.tif', the rasterized and merged soil MUKEY grid in coordinate system of DEM
\item 'soil.tif', a version warped to UTM (and optionally masked/cropped to the catchment)
}

\code{buffer} and \code{threads} determine if and how the UTM grid is masked (as in \code{?save_land})
In a normal workflow you should call \code{get_soil(...)} to write the two source
datasets to disk and make the merged raster, then call \code{save_soil(overwrite=TRUE)}
to save the merged raster in the two files listed above.

See also \code{?save_land}, where \code{buffer} and \code{threads} have the same meaning.
}
\examples{
save_soil('/example')
}
