% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soils.R
\name{save_soils}
\alias{save_soils}
\title{Save the output of \code{get_soils} to disk}
\usage{
save_soils(data_dir, soils = NULL, overwrite = FALSE)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{soils}{SpatRaster of soil MUKey values}

\item{overwrite}{logical if \code{FALSE} the function just returns the file that would be written}
}
\value{
the file name to write
}
\description{
With default \code{overwrite=FALSE} the function returns in a list the three sets
of soils files that are written by this package. If \code{overwrite=TRUE} the one-layer
combined output from STATSGO2 and SSURGO is written to disk.
}
\details{
In a normal workflow you should call \code{get_soils(...)} to write the two source
datasets to disk and generate the merged layer in memory. then call
\code{save_soils(overwrite=TRUE)} to write this merged layer to disk.
}
\examples{
save_soils('/example')
}
