% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_split.R
\name{save_split}
\alias{save_split}
\title{Save results of \code{get_split}}
\usage{
save_split(data_dir, sub_list = NULL, overwrite = FALSE, complete = TRUE)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{sub_list}{list returned from \code{get_split}}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}
}
\value{
the directory paths to write
}
\description{
This saves the list of sub-catchment features returned by \code{get_split} by calling
\code{save_catch} in loop and writing the output to sub-directories of the "split" directory
in \code{data_dir}. Directory names are taken from \code{names(sub_list)}
}
\details{
The function returns a vector of paths to the sub-catchment directories. Pass any of these
paths to \code{save_catch(overwrite=FALSE, extra=TRUE)} to get a list of the files written there.

If no sub-catchment directories are found in the expected location and \code{overwrite=FALSE},
the function returns an empty character.
}
\examples{
save_split('/example')
save_split('/example', extra=TRUE)
}
