% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_soils.R
\name{save_statsgo}
\alias{save_statsgo}
\title{Save the output of \code{get_statsgo} to disk}
\usage{
save_statsgo(data_dir, soils = NULL, model = "statsgo", overwrite = FALSE)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{soils}{sf dataframe returned by \code{get_statsgo}}

\item{model}{character either 'statsgo' or 'ssurgo'}

\item{overwrite}{logical, if \code{TRUE} the function writes to output files if they don't exist}
}
\value{
the file names to write
}
\description{
When \code{overwrite=TRUE} the function writes the data in \code{soils} to a pair of files
named \code{model}, with extensions '.tif' and '.geojson'. When \code{overwrite=FALSE} the
function writes nothing but returns the file paths that would be written.
}
\details{
The function writes a copy of \code{soils} as geoJSON as well as a rasterized version
with grid matching the DEM downloaded in \code{get_dem}. The contents of the "raw"
sub-directory are written by \code{get_statsgo}.
}
\examples{
save_statsgo('/example')
}
