% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_qswat.R
\name{split_catch}
\alias{split_catch}
\title{Split a catchment into sub-catchments with outlet(s) at NWIS gages}
\usage{
split_catch(
  data_dir,
  gage = NULL,
  fast = FALSE,
  snap_dist = units::set_units(100, m),
  s2 = FALSE
)
}
\arguments{
\item{data_dir}{character path to the directory to use for output files}

\item{gage}{sf points data frame, point of interest at which to split the catchment}

\item{fast}{logical if \code{TRUE} only the 'boundary', 'gage', ''}

\item{snap_dist}{numeric with units, distance within which to snap an \code{outlet} point}

\item{s2}{logical if \code{FALSE} sets \code{sf::sf_use_s2(FALSE)} during function evaluation}
}
\value{
something
}
\description{
For more on \code{s2} see \code{?get_upstream}
}
